all changes

+++ src/app/pages/placement/quetions/quetions.component.ts
@@ -3,6 +3,7 @@ import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ToastrService } from 'ngx-toastr';
 import { PlacementService } from 'src/app/core/services/placement.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-quetions',
@@ -315,15 +316,30 @@ export class QuetionsComponent implements OnInit {
     this.modalService.open(exlargeModal, { size: 'lg', windowClass: 'modal-holder', centered: true });
   }
 
-  removeSelfAssessmentQuestionSet(id: any) {
-    this.placementService.removeSelfAssessmentQuestionSet(id).subscribe({
-      next: (res: any) => {
-        this.toastr.success('Question set removed successfully.');
-        this.loadAllData();
-      },
-      error: (err) => {
-        this.toastr.error('Error removing question set.', 'Error');
-        console.error('Error removing question set:', err);
+
+  removeSelfAssessmentQuestionSet(id: string) {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'Do you want to delete this job opening?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.placementService.removeSelfAssessmentQuestionSet(id).subscribe({
+          next: (res: any) => {
+            this.loadAllData();
+            this.getAllSelfQuestionSetDetails();
+            this.toastr.success('Question set removed successfully.');
+          },
+          error: (err) => {
+            this.toastr.error('Error removing question set.', 'Error');
+            console.error('Error removing question set:', err);
+          }
+        });
       }
     });
   }
